Moved Discovery() out of TestContainer.
