+++ test/Beta.Tests/TestContainerTests.cs
@@ -1,56 +1,9 @@
-﻿using System.Diagnostics.CodeAnalysis;
-using Microsoft.Extensions.DependencyInjection;
+﻿using Microsoft.Extensions.DependencyInjection;
 
 namespace Beta.Tests;
 
 public class TestContainerTests
 {
-    public class DiscoverMethod : TestContainerTests
-    {
-        [Fact]
-        public void ReturnsSingleTest()
-        {
-            // Arrange
-            var container = new StubWithTest();
-
-            // Act
-            var tests = container.Discover().ToArray();
-
-            // Assert
-            tests.ShouldHaveSingleItem();
-        }
-
-        [Fact]
-        public void ReturnsNoTests()
-        {
-            // Arrange
-            var container = new StubWithNoTests();
-
-            // Act
-            var tests = container.Discover().ToArray();
-
-            // Assert
-            tests.ShouldBeEmpty();
-        }
-
-        private class StubWithTest : TestContainer
-        {
-            public BetaTest SingleTest()
-            {
-                return Test(() => new Proof<int>(42));
-            }
-        }
-
-        private class StubWithNoTests : TestContainer
-        {
-            [UsedImplicitly]
-            [SuppressMessage("Performance", "CA1822:Mark members as static")]
-            public void NotATest()
-            {
-            }
-        }
-    }
-
     public class PrepareMethod : TestContainerTests
     {
         [Fact]
@@ -129,15 +82,9 @@ public class TestContainerTests
                 services.AddSingleton<StubType>();
             }
 
-            public new Step<T> Require<T>() where T : notnull
-            {
-                return base.Require<T>();
-            }
+            public new Step<T> Require<T>() where T : notnull => base.Require<T>();
 
-            public new Step<object> Require(Type type)
-            {
-                return base.Require(type);
-            }
+            public new Step<object> Require(Type type) => base.Require(type);
         }
     }
 }
